Style: timer 컴포넌트 관심사 분리를 위한 코드 이동 및 커스텀 훅 이름 변경
