+++ src/pages/study-room/hooks/useStepInfo.ts
@@ -1,21 +1,30 @@
-import { useCallback, useEffect, useRef, useState } from 'react';
-import useAudio from './useAudio';
-
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { STUDY_ROOM_STEP } from '@/constant/study-room';
 import { StepInfo } from '@/types/study-room.types';
+import useAudio from './useAudio';
 
 const DELAY = 1000;
 
-const useTimer = ({ step, progressTime, updateAt }: StepInfo) => {
+const useStepInfo = ({ step, progressTime, updateAt }: StepInfo) => {
   const [remainingSeconds, setRemainingSeconds] = useState(0);
 
   const { play } = useAudio();
 
   const intervalRef = useRef<NodeJS.Timeout | null>(null);
 
+  const stepLabel = useMemo(() => STUDY_ROOM_STEP[step], [step]);
+
   const calcRemainingSeconds = useCallback(
     (endTime: number) => Math.ceil((endTime - Date.now()) / 1000),
     [],
   );
+
+  const currentTime = useMemo(() => {
+    const 분 = Math.floor(remainingSeconds / 60);
+    const 초 = remainingSeconds % 60;
+
+    return `${String(분).padStart(2, '0')}:${String(초).padStart(2, '0')}`;
+  }, [remainingSeconds]);
 
   useEffect(() => {
     const startTime = new Date(updateAt).getTime();
@@ -46,7 +55,7 @@ const useTimer = ({ step, progressTime, updateAt }: StepInfo) => {
 
   const clearTimer = () => clearInterval(intervalRef.current as NodeJS.Timeout);
 
-  return { remainingSeconds };
+  return { stepLabel, currentTime };
 };
 
-export default useTimer;
+export default useStepInfo;
